+++ packages/eslint-config-frontline/index.js
@@ -4,8 +4,15 @@ module.exports = {
     ...craEslintConfig,
     // any plugins are okay, except things relating to flow which we dont use
     plugins: [...craEslintConfig.plugins.filter(p => !p.match("flowtype"))],
-    // eslint-config-react-app sets a bunch of TypeScript specific overrides - we dont need them (for now)
-    overrides: [],
+    overrides: [
+        {
+            ...craEslintConfig.overrides[0],
+            parserOptions: {
+                ...craEslintConfig.overrides[0].parserOptions,
+                project: "./tsconfig.json"
+            }
+        }
+    ],
     // any rules are okay, except things relating to flow which we dont use
     rules: Object.keys(craEslintConfig.rules)
         .filter(key => !key.match("flowtype"))
@@ -13,5 +20,11 @@ module.exports = {
             obj[key] = craEslintConfig.rules[key];
             return obj;
         }, {}),
-    extends: ["plugin:prettier/recommended"]
+    extends: ["plugin:prettier/recommended"],
+    settings: {
+        ...craEslintConfig.settings,
+        "import/resolver": {
+            typescript: {}
+        }
+    }
 };