+++ packages/js-config-webpack-plugin/src/config/development.config.ts
@@ -1,3 +1,4 @@
+import ForkTsCheckerWebpackPlugin from "fork-ts-checker-webpack-plugin";
 import { FILE_REGEX, SVG_COMPONENT_REGEX } from "../utils/file-extension";
 import { FrontlineJsConfigWebpackPluginOptions } from "./FrontlineJsConfigWebpackPlugin";
 
@@ -35,5 +36,8 @@ export = (options: FrontlineJsConfigWebpackPluginOptions) => ({
             }
         ]
     },
-    plugins: []
+    plugins: [
+        options.browserslistEnv === "modern" &&
+            new ForkTsCheckerWebpackPlugin({ eslint: false })
+    ].filter(Boolean)
 });