+++ packages/js-config-webpack-plugin/test/fixtures/babel-react/babel.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-    presets: ["@babel/env", "@babel/preset-react"],
+    presets: ["@babel/env", "@babel/preset-react", "@babel/preset-typescript"],
     plugins: ["@babel/plugin-proposal-class-properties"],
     env: {
         testDevelopment: {