+++ packages/font-config-webpack-plugin/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.2-alpha.0](https://github.com/akqa-frontline/frontline/compare/v0.7.1...v0.7.2-alpha.0) (2020-05-22)
+
+**Note:** Version bump only for package @akqa-frontline/font-config-webpack-plugin
+
+
+
+
+
+## [0.7.2-alpha.0](https://github.com/akqa-frontline/frontline/compare/v0.7.2-alpha.2...v0.7.2-alpha.0) (2020-05-22)
+
+**Note:** Version bump only for package @akqa-frontline/font-config-webpack-plugin
+
+
+
+
+
+## [0.7.2-alpha.2](https://github.com/akqa-frontline/frontline/compare/v0.7.2-alpha.1...v0.7.2-alpha.2) (2020-05-22)
+
+**Note:** Version bump only for package @akqa-frontline/font-config-webpack-plugin
+
+
+
+
+
+## [0.7.2-alpha.1](https://github.com/akqa-frontline/frontline/compare/v0.7.2-alpha.0...v0.7.2-alpha.1) (2020-05-22)
+
+**Note:** Version bump only for package @akqa-frontline/font-config-webpack-plugin
+
+
+
+
+
+
 ## [0.7.1](https://github.com/akqa-frontline/frontline/compare/v0.7.0...v0.7.1) (2020-03-03)
 
 **Note:** Version bump only for package @akqa-frontline/font-config-webpack-plugin