+++ packages/webpack-config/CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+
 ## [0.7.2](https://github.com/akqa-frontline/frontline/compare/v0.7.1...v0.7.2) (2020-06-01)